+++ cint/cint/src/auxu.cxx
@@ -21,7 +21,7 @@ int G__readline_FastAlloc(FILE* fp, G__FastAllocString& line, G__FastAllocString
                           int* argn, char* arg[])
 {
    // -- FIXME: Describe this function!
-   char* null_fgets = fgets(line, line.Capacity() - 1, fp);
+   char* null_fgets = fgets(line, (size_t) (line.Capacity() - 1), fp);
    if (null_fgets) {
       argbuf = line;
       G__split(line, argbuf, argn, arg);
