+++ cint/cint/src/scrupto.cxx
@@ -294,7 +294,7 @@ static void G__close_inputfiles_upto(G__dictposition* pos)
       if (G__srcfile[G__nfile].filename) {
          // --
 #ifndef G__OLDIMPLEMENTATION1546
-         unsigned int len = strlen(G__srcfile[G__nfile].filename);
+         size_t len = strlen(G__srcfile[G__nfile].filename);
          if (
             (len > strlen(G__NAMEDMACROEXT2)) &&
             !strcmp(G__srcfile[G__nfile].filename + len - strlen(G__NAMEDMACROEXT2), G__NAMEDMACROEXT2)
@@ -643,11 +643,11 @@ static int G__destroy_upto_vararray(G__var_array* var, int global, int ig15)
                   // FIXME: Do we really need this special case?
                   long store_globalvarpointer = G__globalvarpointer;
                   int size = G__struct.size[G__tagnum];
-                  int num_of_elements = var->varlabel[idx][1] /* number of elements */;
+                  size_t num_of_elements = var->varlabel[idx][1] /* number of elements */;
                   if (!num_of_elements) {
                      num_of_elements = 1;
                   }
-                  for (int i = num_of_elements - 1; i >= 0; --i) {
+                  for (long i = num_of_elements - 1; i >= 0; --i) {
                      G__store_struct_offset = var->p[idx] + (i * size);
                      G__globalvarpointer = G__store_struct_offset;
                      int known = 0;
@@ -664,7 +664,7 @@ static int G__destroy_upto_vararray(G__var_array* var, int global, int ig15)
                else {
                   // -- The variable is *not* and array with auto storage duration.
                   G__store_struct_offset = var->p[idx];
-                  int i = var->varlabel[idx][1] /* number of elements */;
+                  size_t i = var->varlabel[idx][1] /* number of elements */;
                   if (i > 0)  {
                      G__cpp_aryconstruct = i;
                   }
@@ -676,11 +676,11 @@ static int G__destroy_upto_vararray(G__var_array* var, int global, int ig15)
             else {
                // -- The class is interpreted.
                int size = G__struct.size[G__tagnum];
-               int num_of_elements = var->varlabel[idx][1] /* number of elements */;
+               size_t num_of_elements = var->varlabel[idx][1] /* number of elements */;
                if (!num_of_elements) {
                   num_of_elements = 1;
                }
-               for (int i = num_of_elements - 1; i >= 0; --i) {
+               for (long i = num_of_elements - 1; i >= 0; --i) {
                   G__store_struct_offset = var->p[idx] + (i * size);
                   if (G__dispsource) {
                      G__fprinterr(G__serr, "\n0x%lx.%s", G__store_struct_offset, temp());
@@ -710,7 +710,7 @@ static int G__destroy_upto_vararray(G__var_array* var, int global, int ig15)
             var->p[idx] // has memory allocated.
          ) {
             // -- Decrement the reference counts for each member of the pointer array.
-            int i = var->varlabel[idx][1] /* number of elements */;
+            long i = var->varlabel[idx][1] /* number of elements */;
             if (!i) {
                i = 1;
             }
@@ -1204,7 +1204,7 @@ int G__close_inputfiles()
       if (G__srcfile[iarg].filename) {
          // --
 #ifndef G__OLDIMPLEMENTATION1546
-         int len = strlen(G__srcfile[iarg].filename);
+         size_t len = strlen(G__srcfile[iarg].filename);
          if (len > (int)strlen(G__NAMEDMACROEXT2) &&
                strcmp(G__srcfile[iarg].filename + len - strlen(G__NAMEDMACROEXT2),
                       G__NAMEDMACROEXT2) == 0) {
